Add integration test, lint, reorganize code, fix defaults
