+++ modules/frontend/frontend_test.go
@@ -1,8 +1,9 @@
 package frontend
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestCreateBlockShards(t *testing.T) {
